When comparing, cast to the larger size, off_t in this case. Once we know which one is smaller, then we cast to the smaller size.
Thanks to Xin Li (delphij@)
Pointy hat: /me


git-svn-id: svn+ssh://svn.freebsd.org/base/head@184669 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f

+++ usr.bin/tar/test/main.c
@@ -494,7 +494,7 @@ test_assert_empty_file(const char *f1fmt, ...)
 	if (fd < 0) {
 		fprintf(stderr, "    Unable to open %s\n", f1);
 	} else {
-		s = (sizeof(buff) < (size_t)st.st_size) ?
+		s = ((off_t)sizeof(buff) < st.st_size) ?
 		    (ssize_t)sizeof(buff) : (ssize_t)st.st_size;
 		s = read(fd, buff, s);
 		hexdump(buff, NULL, s, 0);
